requirements.txt
gradio>=5.0.0
langchain>=1.0.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-mcp-adapters>=0.1.0
langgraph>=0.2.0
python-dotenv>=1.0.0
duckduckgo-search>=6.0.0
mcp>=1.0.0
nest-asyncio>=1.6.0
requests>=2.31.0
